Small fixes to Java basics
